cambios x2
se agrego decimales en la cantidad, facturacion

+++ ferreteriaVillamil-main/ferreteriaVillamil-main/backend/models/detalle_pedido.js
@@ -4,7 +4,7 @@ module.exports = (sequelize, DataTypes) => {
     id_detalle: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
     id_pedido: DataTypes.INTEGER,
     id_articulo: DataTypes.INTEGER,
-    cantidad: DataTypes.INTEGER,
+    cantidad: DataTypes.DECIMAL(10, 2),
     precio_unitario: DataTypes.DECIMAL,
     subtotal: DataTypes.DECIMAL
   }, {
